+++ ndd/nsb.py
@@ -1,168 +1,281 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2016,2017 Simone Marsili
-# All rights reserved.
+# Author: Simone Marsili <simomarsili@gmail.com>
 # License: BSD 3 clause
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
-from builtins import (  # pylint: disable=redefined-builtin, unused-import
-    bytes, dict, int, list, object, range, str,
-    ascii, chr, hex, input, next, oct, open,
-    pow, round, super,
-    filter, map, zip)
-
-__copyright__ = "Copyright (C) 2016,2017 Simone Marsili"
-__license__ = "BSD 3 clause"
-__author__ = "Simone Marsili (simomarsili@gmail.com)"
-__all__ = ['entropy', 'histogram']
-
+"""Functions module."""
+import logging
 import numpy
+import ndd
+from ndd.estimators import Entropy, JSDivergence
+from ndd.exceptions import (NumericError, HistogramError, AxisError,
+                            CardinalityError)
 
-def _select_estimator(k, alpha, plugin):
-    from ndd import _nsb
+__all__ = ['entropy',
+           'jensen_shannon_divergence',
+           'histogram', ]
 
-    if plugin:
-        if alpha is None:
-            return _nsb.plugin
-        else:
-            return _nsb.pseudo
-    else:
-        if alpha is None:
-            return _nsb.nsb
-        else:
-            #TODO: compute variance over the posterior at fixed alpha
-            return _nsb.dirichlet
+logger = logging.getLogger(__name__)
 
-def entropy(counts, k=None, alpha=None, return_std=False, plugin=False):
+
+def entropy(pk, k=None, alpha=None, plugin=False, return_std=False):
     """
+    Return a Bayesian estimate of the entropy from an array of counts.
+
     Return a Bayesian estimate of the entropy of an unknown discrete
-    distribution from an input array of counts. The estimator relies on a
-    mixture of (properly weighted) Dirichlet priors
-    (Nemenman-Shafee-Bialek estimator).
+    distribution from an input array of counts pk.
 
     Parameters
     ----------
-
-    counts : array_like
-        The number of occurrences of a set of states/classes.
-        The estimate is computed over the flattened array.
-
-    k : int, optional
-        Total number of classes. k >= len(counts).
-        Defaults to len(counts).
-
+    pk : array-like
+        The number of occurrences of a set of bins.
+    k : int or array-like, optional
+        Total number of bins (taking into account unobserved bins);
+        k >= len(pk). A float is a valid input for whole numbers (e.g. k=1.e3).
+        If an array, set k = numpy.prod(k).
+        Defaults to n_bins.
     alpha : float, optional
-        If alpha is passed, use a single Dirichlet prior with concentration
+        If alpha is not None, use a single Dirichlet prior with concentration
         parameter alpha (fixed alpha estimator). alpha > 0.0.
-
-    return_std : boolean, optional
-        If True, return the tuple (estimate, std) where std in an estimate
-        of the standard deviation over the posterior for the entropy
-        of the distribution.
-
     plugin : boolean, optional
-        If True, estimate the distribution over states/classes from the
-        frequencies and then plug the estimate into the entropy definition
-        (plugin estimator).
+        If True, return a 'plugin' estimate of the entropy. The discrete
+        distribution is estimated from the empirical frequencies over bins
+        and inserted into the entropy definition (plugin estimator).
         If alpha is passed in combination with plugin=True, add
         alpha pseudocounts to each frequency count (pseudocount estimator).
+    return_std : boolean, optional
+        If True, also return an approximated value for the standard deviation
+        over the entropy posterior.
 
     Returns
     -------
     entropy : float
         Entropy estimate.
-
     std : float, optional
-        If return_std == True, return an approximation for the standard
-        deviation over the posterior for the entropy.
+        Uncertainty in the entropy estimate
+        (approximated standard deviation over the entropy posterior).
+        Only if `return_std` is True.
+
+    Raises
+    ------
+    NumericError
+        If result is NaN
 
     """
-    from ndd import _nsb
 
-    args_dict = {
-        _nsb.plugin: (),
-        _nsb.pseudo: (k, alpha),
-        _nsb.nsb: (k, ),
-        _nsb.dirichlet: (k, alpha)
-    }
+    # pk is an array of counts
+    estimator = Entropy(alpha, plugin).fit(pk, k)
+    S, err = estimator.estimate_, estimator.err_
 
-    try:
-        counts = numpy.array(counts, dtype=numpy.int32)
-    except ValueError:
-        raise
-    if numpy.any(counts < 0):
-        raise ValueError("Frequency counts cant be negative")
-    # flatten the input array
-    counts = counts.flatten()
+    if numpy.isnan(S):
+        raise NumericError("NaN value")
 
-    n_bins = len(counts)
-    if k is None:
-        k = numpy.int32(n_bins)
+    if return_std:
+        if err is not None and numpy.isnan(err):
+            raise NumericError("NaN value")
+        return S, err
     else:
-        try:
-            k = numpy.int32(k)
-        except ValueError:
-            raise
-        if k < n_bins:
-            raise ValueError("k (%s) is smaller than the number of bins (%s)"
-                             % (k, n_bins))
+        return S
 
-    if alpha:
-        try:
-            alpha = numpy.float64(alpha)
-        except ValueError:
-            raise
-        if alpha <= 0:
-            raise ValueError("alpha <= 0")
 
-        
-    estimator = _select_estimator(k, alpha, plugin)
+def jensen_shannon_divergence(pk, k=None, alpha=None, plugin=False):
+    """
+    Return the Jensen-Shannon divergence from a matrix of counts.
 
-    if k == 1: # if the total number of classes is one
-        if estimator == _nsb.nsb and return_std:
-            return (0.0, 0.0)
-        else:
-            return 0.0
-
-    args = args_dict[estimator]
-    result = estimator(counts, *args)
-
-    if numpy.any(numpy.isnan(numpy.squeeze(result))):
-        raise FloatingPointError("NaN value")
-
-    if estimator == _nsb.nsb:
-        if not return_std:
-            result = result[0]
-
-    return result
-
-def histogram(data, return_unique=False):
-    """Compute an histogram from data. Wrapper to numpy.unique.
+    Return an estimate of the Jensen-Shannon divergence between
+    n_distributions unknown discrete distributions from a
+    n_distributions-by-n_bins input array of counts.
+    The estimate (in nats) is computed as a combination of single Bayesian
+    entropy estimates. If the total number of samples varies among the
+    distributions, the function returns the divergence between the
+    distributions with weights proportional to the total number of samples in
+    each row (see the general definition of Jensen-Shannon divergence:
+    https://en.wikipedia.org/wiki/Jensen-Shannon_divergence).
 
     Parameters
     ----------
 
-    data : array_like
-        Input data array. If n-dimensional, statistics is computed along
-        axis 0.
-
-    return_unique : bool, optional
-        If True, also return the unique elements corresponding to each bin.
+    pk : array-like, shape (n_distributions, n_bins)
+        Matrix of frequency counts. Each row corresponds to the number of
+        occurrences of a set of bins from a different distribution.
+    k : int or array-like, optional
+        Total number of bins (taking into account unobserved bins); k >= n_bins
+        A float is a valid input for whole numbers (e.g. k=1.e3).
+        If an array, set k = numpy.prod(k).
+        Defaults to n_bins.
+    alpha : float, optional
+        If not None, the entropy estimator uses a single Dirichlet prior with
+        concentration parameter alpha (fixed alpha estimator). alpha > 0.0.
+    plugin : boolean, optional
+        If True, use a 'plugin' estimator for the entropy.
+        If alpha is passed in combination with plugin == True, add alpha
+        pseudoconts to the frequency counts in the plugin estimate.
 
     Returns
     -------
+    float
+        Jensen-Shannon divergence.
 
-    counts : ndarray
-        Bin counts.
-
-    unique : ndarray, optional
-        Unique elements corresponding to each bin in counts.
-        Only if return_elements == True
+    Raises
+    ------
+    NumericError
+        If result is NaN
 
     """
 
-    unique, counts = numpy.unique(data, axis=0, return_counts=True)
+    estimator = JSDivergence(alpha, plugin).fit(pk, k)
+    js = estimator.estimate_
 
-    if return_unique:
-        return (unique, counts)
-    else:
+    if numpy.isnan(js):
+        raise NumericError("NaN value")
+
+    return js
+
+
+def _nbins(data):
+    """
+    The number of unique elements along axis 0. If data is p-dimensional,
+    the num. of unique elements for each variable.
+    """
+    # reshape as a p-by-n array
+    data = as_data_array(data)
+    return [len(numpy.unique(v)) for v in data]
+
+
+def histogram(data, axis=0, r=0):
+    """Compute an histogram from data. Wrapper to numpy.unique.
+
+    Parameters
+    ----------
+    data : array-like
+        An array of n samples from p variables.
+    axis : int, optional
+        The sample-indexing axis
+    r : int, optional
+        If r > 0, return a generator that yields a bin counts array
+        for each possible combination of r variables.
+
+    Returns
+    -------
+    counts : ndarray
+        Bin counts.
+
+    Raises
+    ------
+    HistogramError
+        If r > p.
+
+    """
+    from itertools import combinations
+    # reshape as a p-by-n array
+    data = as_data_array(data, axis=axis)
+    p = data.shape[0]
+    if r > p:
+        raise HistogramError(
+            'r (%r) is larger than the number of variables (%r)' % (r, p))
+    if r == 0:
+        # statistics for the p-dimensional variable
+        _, counts = numpy.unique(data, return_counts=True, axis=1)
         return counts
+    else:
+        return (ndd.histogram(d, axis=1) for d in combinations(data, r=r))
+
+
+def as_data_array(ar, axis=0):
+    """
+    For a generic ndarray, flatten the subarrays indexed by axis `axis`
+    before transposing.
+
+    Attributes
+    ----------
+    ar : array-like
+        Data array.
+    axis : int, optional
+        The sample-indexing axis
+
+    Returns
+    -------
+    transpesed_array
+        2D data array with different samples in different columns.
+
+    Raises
+    ------
+    AxisError
+        axis is invalid
+
+    """
+
+    ar = numpy.asanyarray(ar)
+
+    if ar.ndim == 1:
+        n = ar.shape[0]
+        ar = ar.reshape(1, n)
+    elif ar.ndim == 2:
+        if axis == 0:
+            ar = ar.T
+    elif ar.ndim > 2:
+        if axis != 0:
+            try:
+                ar = numpy.swapaxes(ar, axis, 0)
+            except ValueError:
+                raise AxisError(axis, ar.ndim)
+        n = ar.shape[0]
+        ar = ar.reshape(n, -1)
+        ar = ar.T
+
+    return numpy.ascontiguousarray(ar)
+
+
+def _from_data(ar, ks=None, axis=0, r=0):
+    """
+    Given an array of data, return an entropy estimate.
+
+    Paramaters
+    ----------
+    ar : array-like
+        n-by-p array of n samples from p discrete variables.
+    ks : 1D p-dimensional array, optional
+        Alphabet size for each variable.
+    axis : int, optional
+        The sample-indexing axis.
+    r : int, optional
+        If r > 0, return a generator yielding estimates for the p-choose-r
+        possible combinations of length r from the op variables.
+
+    Returns
+    -------
+    float
+        Entropy estimate
+
+    Raises
+    ------
+    CardinalityError
+        len(ks) != p
+
+    """
+    from itertools import combinations
+
+    ar = as_data_array(ar, axis=axis)
+    p = ar.shape[0]
+
+    if ks is None:
+        ks = numpy.array([len(numpy.unique(v)) for v in ar])
+    else:
+        try:
+            if len(ks) == p:
+                ks = numpy.array(ks)
+            else:
+                raise CardinalityError("k should have len %s" % p)
+        except TypeError as e:
+            raise CardinalityError(e)
+
+    entropy_estimator = Entropy()
+    if r == 0:
+        counts = histogram(ar, axis=1)
+        k = numpy.prod(ks)
+        return entropy_estimator(counts, k=k)
+    else:
+        counts_combinations = histogram(ar, axis=1, r=r)
+        alphabet_size_combinations = (numpy.prod(x)
+                                      for x in combinations(ks, r=r))
+        return (
+            entropy_estimator(c, k=k)
+            for c, k in zip(counts_combinations, alphabet_size_combinations))