+++ tests/make_test_ref.py
@@ -5,44 +5,37 @@ import itertools
 import json
 import ndd
 
+
 def random_counts(n=None, k=None, alpha=None):
     random.seed(123)
     pp = random.dirichlet([alpha]*k)
     return random.multinomial(n, pp)
 
+
 # a test should include:
-# name, input setting, ndd setting, result
+# input setting, ndd setting, result
+prms = [
+    {},  # NSB
+    {'alpha': 1.0},  # Dirichlet
+    {'alpha': 0.0, 'plugin': 1},  # ML
+    {'alpha': 1.0, 'plugin': 1},  # pseudo
+]
 
-methods = ['NSB', 'Dirichlet', 'ML', 'pseudo']
-
-ndd_kwargs = {
-    'NSB' : {},
-    'Dirichlet' : {'alpha':1.0},
-    'ML' : {'alpha':0.0, 'plugin':1},
-    'pseudo' : {'alpha':1.0, 'plugin':1},
-}
-
-n_array = [int(x) for x in numpy.logspace(1, 3, num=3)]
-k_array = [int(x) for x in numpy.logspace(1, 3, num=3)]
-a_array = numpy.logspace(-2, 1, num=4)
+n_vals = [int(x) for x in numpy.logspace(1, 3, num=3)]
+k_vals = [int(x) for x in numpy.logspace(1, 3, num=3)]
+a_vals = numpy.logspace(-2, 1, num=4)
 
 # list of combinations of parameter values
-settings = [{'n':n, 'k':k, 'alpha':alpha}
-            for n, k, alpha in itertools.product(n_array, k_array, a_array)]
+settings = ({'n': n, 'k': k, 'alpha': alpha}
+            for n, k, alpha in itertools.product(n_vals, k_vals, a_vals))
 
-test_methods = {}
-for method in methods:
-    tests = []
-    kwargs = ndd_kwargs[method]
-    for setting in settings:
-        kwargs['k'] = setting['k']
-        counts = random_counts(**setting)
-        result = ndd.entropy(counts, **kwargs)
-        ndd_in = {'k':setting['k']}
-        for key, val in kwargs.items():
-            ndd_in[key] = val
-        test = (setting, ndd_in, result)
-        tests.append(test)
-    test_methods[method] = tests
-json.dump(test_methods, sys.stdout)
-
+tests = []
+for setting in settings:
+    k = setting['k']
+    counts = random_counts(**setting)
+    for kwargs in prms:
+        # n. of classes is known
+        result = ndd.entropy(counts, **kwargs, k=k)
+        test_case = (setting, kwargs, result)
+        tests.append(test_case)
+json.dump(tests, sys.stdout)